LayoutTests/fast/events/mac/pointer-event-uses-platform-event-timestamp.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
<meta charset="utf-8">
<meta name="viewport" content="width=device-width, initial-scale=1">
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/js-test.js"></script>
<style>
body, html {
    margin: 0;
}

.box {
    width: 100px;
    height: 100px;
    background: tomato;
    position: fixed;
    top: 0;
    left: 0;
    z-index: 100;
}
</style>
</head>
<body>
<div class="box"></div>
</body>
<script>
jsTestIsAsync = true;

addEventListener("load", async (loadEvent) => {
    pointerDownTimestamp = -1;
    mouseDownTimestamp = 0;

    document.body.addEventListener('mousedown', (event) => {
        mouseDownTimestamp = event.timeStamp;
    });

    document.body.addEventListener('pointerdown', (event) => {
        pointerDownTimestamp = event.timeStamp;
    });

    await UIHelper.activateAt(50, 50);

    shouldBecomeEqual("mouseDownTimestamp", "pointerDownTimestamp", finishJSTest);
});
</script>
</html>
